+++ types/index.d.ts
@@ -1,2 +1,3 @@
 export * as JsonTypes from "./jsonTypes";
 export * as Compiler from "./compiler/index";
+export { Var } from "./compiler/utils";